# apps/marketplace/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Marketplace Simulator API
  version: 1.0.0
  description: |
    A provider-aware marketplace API. Order and fulfillment contracts are only
    available through `SHOPEE_LIKE` and `TOKOPEDIA_LIKE`; Generic is not an
    order-provider profile.

    Shared catalogue, warehouse, and webhook requests under `/api/v1` are HMAC-SHA256 signed using `METHOD + PATH + UNIX_TIMESTAMP + RAW_BODY`.
    Every mutating public operation across the shared, Shopee-like, and
    Tokopedia-like contracts require an `Idempotency-Key`. Concurrent use is
    claimed atomically, completed responses are replayed, and reusing a key for
    a different logical request returns `409`.

    Successful and rate-limited public responses expose `X-RateLimit-Limit`,
    `X-RateLimit-Remaining`, and `X-RateLimit-Reset` (Unix seconds). Public requests can return
    `503 MARKETPLACE_MAINTENANCE` when the administrator enables global maintenance mode.

    Webhooks are asynchronous. A 2xx response succeeds; all other responses and network failures
    retry at 30 seconds, 2 minutes, 10 minutes, and 30 minutes. Delivery attempts are immutable.
    Per-shop scenarios can inject slow responses, random 500s, timeouts, rate limits, duplicate,
    delayed, out-of-order, and failed webhook deliveries without affecting other shops.

    `SHOPEE_LIKE` shops additionally expose `/api/shopee/v1`. This is a distinct
...
    `X-Shopee-Signature`, where the signature is HMAC-SHA256 of
...

...
servers:
...
paths:
...
  /api/shopee/v1/webhooks:
    get:
      operationId: shopeeListWebhooks
      tags: [Webhooks]
      summary: List provider-style webhook registrations
      parameters: [{ $ref: '#/components/parameters/ShopeePartnerID' }, { $ref: '#/components/parameters/ShopeeTimestamp' }, { $ref: '#/components/parameters/ShopeeSignature' }]
      responses: { '200': { $ref: '#/components/responses/ShopeeResponse' }, '401': { $ref: '#/components/responses/ShopeeError' } }
    post:
      operationId: shopeeCreateWebhook
      tags: [Webhooks]
      summary: Register callback_url for item_update, order_status_update, or logistics_status_update
      parameters: [{ $ref: '#/components/parameters/ShopeePartnerID' }, { $ref: '#/components/parameters/ShopeeTimestamp' }, { $ref: '#/components/parameters/ShopeeSignature' }, { $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [callback_url, event_types], properties: { callback_url: { type: string, format: uri }, secret: { type: string }, event_types: { type: array, items: { type: string, enum: [item_update, order_status_update, logistics_status_update] } } } } } } }
      responses: { '200': { $ref: '#/components/responses/ShopeeResponse' }, '400': { $ref: '#/components/responses/ShopeeError' }, '401': { $ref: '#/components/responses/ShopeeError' }, '409': { $ref: '#/components/responses/ShopeeError' }, '413': { $ref: '#/components/responses/ShopeeError' } }
...
  /api/tokopedia/v202309/orders/{id}:
    parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
    get:
      operationId: tokopediaGetOrder
      tags: [Tokopedia Like]
      summary: Get a TOKOPEDIA_LIKE order and its package/shipment reference
      parameters: [{ $ref: '#/components/parameters/TokopediaAppKey' }, { $ref: '#/components/parameters/TokopediaTimestamp' }, { $ref: '#/components/parameters/TokopediaSign' }, { $ref: '#/components/parameters/TokopediaAccessToken' }]
      responses: { '200': { $ref: '#/components/responses/TokopediaResponse' }, '400': { $ref: '#/components/responses/TokopediaError' }, '429': { $ref: '#/components/responses/TokopediaError' } }
  /api/tokopedia/v202309/orders/{id}/pack:
    post:
      operationId: tokopediaPackOrder
      tags: [Tokopedia Like]
      summary: Pack a paid order into seller fulfillment (`ON_HOLD` to `AWAITING_SHIPMENT`)
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }, { $ref: '#/components/parameters/TokopediaAppKey' }, { $ref: '#/components/parameters/TokopediaTimestamp' }, { $ref: '#/components/parameters/TokopediaSign' }, { $ref: '#/components/parameters/TokopediaAccessToken' }, { $ref: '#/components/parameters/IdempotencyKey' }]
      responses: { '200': { $ref: '#/components/responses/TokopediaResponse' }, '400': { $ref: '#/components/responses/TokopediaError' }, '401': { $ref: '#/components/responses/TokopediaError' }, '409': { $ref: '#/components/responses/TokopediaError' }, '413': { $ref: '#/components/responses/TokopediaError' }, '429': { $ref: '#/components/responses/TokopediaError' } }
  /api/tokopedia/v202309/orders/{id}/handover:
    post:
      operationId: tokopediaHandoverOrder
      tags: [Tokopedia Like]
      summary: Hand over a packed order (`AWAITING_SHIPMENT` to `AWAITING_COLLECTION`)
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }, { $ref: '#/components/parameters/TokopediaAppKey' }, { $ref: '#/components/parameters/TokopediaTimestamp' }, { $ref: '#/components/parameters/TokopediaSign' }, { $ref: '#/components/parameters/TokopediaAccessToken' }, { $ref: '#/components/parameters/IdempotencyKey' }]
      responses: { '200': { $ref: '#/components/responses/TokopediaResponse' }, '400': { $ref: '#/components/responses/TokopediaError' }, '401': { $ref: '#/components/responses/TokopediaError' }, '409': { $ref: '#/components/responses/TokopediaError' }, '413': { $ref: '#/components/responses/TokopediaError' }, '429': { $ref: '#/components/responses/TokopediaError' } }
...
  /api/tokopedia/v202309/webhooks:
    put:
      operationId: tokopediaConfigureWebhooks
      tags: [Tokopedia Like, Webhooks]
      summary: Configure a TOKOPEDIA_LIKE callback for order, package, or product notifications
      parameters: [{ $ref: '#/components/parameters/TokopediaAppKey' }, { $ref: '#/components/parameters/TokopediaTimestamp' }, { $ref: '#/components/parameters/TokopediaSign' }, { $ref: '#/components/parameters/TokopediaAccessToken' }, { $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [callback_url, event_types], properties: { callback_url: { type: string, format: uri }, secret: { type: string, description: Callback verification secret. }, event_types: { type: array, items: { type: string, enum: [ORDER_STATUS_CHANGE, PACKAGE_UPDATE, PRODUCT_INFORMATION_CHANGE] } } } } } } }
      responses: { '200': { $ref: '#/components/responses/TokopediaResponse' }, '400': { $ref: '#/components/responses/TokopediaError' }, '401': { $ref: '#/components/responses/TokopediaError' }, '409': { $ref: '#/components/responses/TokopediaError' }, '413': { $ref: '#/components/responses/TokopediaError' }, '429': { $ref: '#/components/responses/TokopediaError' } }
...
  /api/v1/webhooks:
    get:
      operationId: listWebhooks
      tags: [Webhooks]
      summary: List webhook registrations
      parameters: [{ $ref: '#/components/parameters/ClientID' }, { $ref: '#/components/parameters/Timestamp' }, { $ref: '#/components/parameters/Signature' }]
      responses: { '200': { $ref: '#/components/responses/RateLimitedWebhookList' }, '401': { $ref: '#/components/responses/Error' }, '429': { $ref: '#/components/responses/RateLimited' }, '503': { $ref: '#/components/responses/Maintenance' } }
    post:
      operationId: createWebhook
      tags: [Webhooks]
      summary: Register a webhook; omit secret to have one generated and returned once
      parameters: [{ $ref: '#/components/parameters/ClientID' }, { $ref: '#/components/parameters/Timestamp' }, { $ref: '#/components/parameters/Signature' }, { $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [url, subscribed_events], properties: { url: { type: string, format: uri }, secret: { type: string, writeOnly: true }, subscribed_events: { type: array, items: { type: string, enum: [product.created, product.updated, product.deleted, order.created, order.paid, order.payment_failed, order.payment_expired, order.processing, order.ready_to_ship, order.shipped, order.in_delivery, order.delivered, order.completed, order.cancelled, order.sla_expired] } } } } } } }
      responses: { '201': { $ref: '#/components/responses/RateLimitedWebhookRegistration' }, '400': { $ref: '#/components/responses/Error' }, '401': { $ref: '#/components/responses/Error' }, '409': { $ref: '#/components/responses/Error' }, '413': { $ref: '#/components/responses/Error' }, '429': { $ref: '#/components/responses/RateLimited' }, '503': { $ref: '#/components/responses/Maintenance' } }
  /api/v1/webhooks/{id}:
    delete:
      operationId: deleteWebhook
      tags: [Webhooks]
      summary: Delete a webhook registration
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }, { $ref: '#/components/parameters/ClientID' }, { $ref: '#/components/parameters/Timestamp' }, { $ref: '#/components/parameters/Signature' }, { $ref: '#/components/parameters/IdempotencyKey' }]
      responses: { '204': { description: Webhook deleted, headers: { X-RateLimit-Limit: { $ref: '#/components/headers/RateLimitLimit' }, X-RateLimit-Remaining: { $ref: '#/components/headers/RateLimitRemaining' }, X-RateLimit-Reset: { $ref: '#/components/headers/RateLimitReset' } } }, '401': { $ref: '#/components/responses/Error' }, '404': { $ref: '#/components/responses/Error' }, '409': { $ref: '#/components/responses/Error' }, '413': { $ref: '#/components/responses/Error' }, '429': { $ref: '#/components/responses/RateLimited' }, '503': { $ref: '#/components/responses/Maintenance' } }
...
components:
  securitySchemes:
    ControlPlaneBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque-session-token
  parameters:
    ClientID: { name: X-Client-Id, in: header, required: true, schema: { type: string } }
    Timestamp: { name: X-Timestamp, in: header, required: true, schema: { type: string, description: Unix seconds } }
    Signature: { name: X-Signature, in: header, required: true, schema: { type: string, description: Hex HMAC-SHA256 of METHOD + PATH + TIMESTAMP + raw body } }
    ShopeePartnerID: { name: X-Shopee-Partner-Id, in: header, required: true, schema: { type: string }, description: Credential client ID used as the provider partner ID. }
    ShopeeTimestamp: { name: X-Shopee-Timestamp, in: header, required: true, schema: { type: string, description: Unix seconds. } }
    ShopeeSignature: { name: X-Shopee-Signature, in: header, required: true, schema: { type: string, description: Hex HMAC-SHA256 of PARTNER_ID + PATH + TIMESTAMP + raw body. } }
    TokopediaAppKey: { name: app_key, in: query, required: true, schema: { type: string }, description: Partner Center app key (the control-plane credential client ID). }
    TokopediaTimestamp: { name: timestamp, in: query, required: true, schema: { type: integer, format: int64 }, description: Unix seconds; accepted within five minutes. }
    TokopediaSign: { name: sign, in: query, required: true, schema: { type: string }, description: Hex HMAC-SHA256 of APP_SECRET + PATH + sorted query excluding sign/access_token + raw body + APP_SECRET. }
    TokopediaAccessToken: { name: x-tts-access-token, in: header, required: true, schema: { type: string }, description: Access token returned only when the control-plane credential is created. }
    IdempotencyKey: { name: Idempotency-Key, in: header, required: true, schema: { type: string, maxLength: 255 }, description: Stable key for one logical mutation. Retrying the identical method, path, stable query, and body replays its completed response; a different payload returns 409. }
...
  headers:
    RateLimitLimit: { description: Configured requests allowed in the current minute for this credential., schema: { type: integer } }
    RateLimitRemaining: { description: Remaining requests in the current minute. Zero on a 429 response., schema: { type: integer, minimum: 0 } }
    RateLimitReset: { description: Unix timestamp when the quota window resets., schema: { type: integer, format: int64 } }
  schemas:
...
    WebhookRegistration:
      type: object
      required: [id, shop_id, url, enabled, subscribed_events]
      properties:
        id: { type: string }
        shop_id: { type: string }
        url: { type: string, format: uri }
        enabled: { type: boolean }
        subscribed_events: { type: array, items: { type: string, enum: [product.created, product.updated, product.deleted, order.created, order.paid, order.processing, order.ready_to_ship, order.shipped, order.in_delivery, order.delivered, order.completed, order.cancelled] } }
        secret: { type: string, writeOnly: true, description: Returned only once when the simulator generated it. }
    WebhookList:
      type: object
      required: [data]
      properties:
        data: { type: array, items: { $ref: '#/components/schemas/WebhookRegistration' } }
    Error:
      type: object
      properties: { error: { type: object, properties: { code: { type: string }, message: { type: string } } } }
  responses:
    ShopeeResponse: { description: Successful SHOPEE_LIKE response, headers: { X-Shopee-Api-Call-Limit: { schema: { type: string }, description: Calls used/current limit. }, X-Shopee-RateLimit-Reset: { schema: { type: integer }, description: Unix reset time. } }, content: { application/json: { schema: { type: object, required: [error, message, request_id, response], properties: { error: { type: string, example: '' }, message: { type: string, example: success }, request_id: { type: string }, response: { type: object } } } } } }
    ShopeeError: { description: SHOPEE_LIKE error, content: { application/json: { schema: { type: object, required: [error, message, request_id], properties: { error: { type: string, example: error_invalid_state }, message: { type: string }, request_id: { type: string } } } } } }
    TokopediaResponse: { description: Successful TOKOPEDIA_LIKE response, headers: { X-TTS-Api-Call-Limit: { schema: { type: integer } }, X-TTS-RateLimit-Reset: { schema: { type: integer, description: Unix reset time. } } }, content: { application/json: { schema: { type: object, required: [code, message, request_id, data], properties: { code: { type: integer, example: 0 }, message: { type: string, example: Success }, request_id: { type: string }, data: { type: object } } } } } }
    TokopediaError: { description: TOKOPEDIA_LIKE error, content: { application/json: { schema: { type: object, required: [code, message, request_id, data], properties: { code: { type: integer, example: 36000003 }, message: { type: string }, request_id: { type: string }, data: { type: object } } } } } }
    Error: { description: API error, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    Maintenance: { description: Global maintenance is enabled., content: { application/json: { schema: { $ref: '#/components/schemas/Error' }, example: { error: { code: MARKETPLACE_MAINTENANCE, message: marketplace is under maintenance } } } } }
    RateLimited: { description: Credential quota is exhausted or the per-shop rate-limit scenario is enabled., headers: { X-RateLimit-Limit: { $ref: '#/components/headers/RateLimitLimit' }, X-RateLimit-Remaining: { $ref: '#/components/headers/RateLimitRemaining' }, X-RateLimit-Reset: { $ref: '#/components/headers/RateLimitReset' } }, content: { application/json: { schema: { $ref: '#/components/schemas/Error' }, example: { error: { code: RATE_LIMIT_EXCEEDED, message: rate limit exceeded } } } } }
...
    RateLimitedWebhookList: { description: Webhook list with credential quota headers., headers: { X-RateLimit-Limit: { $ref: '#/components/headers/RateLimitLimit' }, X-RateLimit-Remaining: { $ref: '#/components/headers/RateLimitRemaining' }, X-RateLimit-Reset: { $ref: '#/components/headers/RateLimitReset' } }, content: { application/json: { schema: { $ref: '#/components/schemas/WebhookList' } } } }
    RateLimitedWebhookRegistration: { description: Webhook registration with credential quota headers. A generated secret is returned only by this create response., headers: { X-RateLimit-Limit: { $ref: '#/components/headers/RateLimitLimit' }, X-RateLimit-Remaining: { $ref: '#/components/headers/RateLimitRemaining' }, X-RateLimit-Reset: { $ref: '#/components/headers/RateLimitReset' } }, content: { application/json: { schema: { $ref: '#/components/schemas/WebhookRegistration' } } } }